Fix accidental failure to override virtual method
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9215 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/cxxtest/include/cxxtest/Gui.h
@@ -70,7 +70,7 @@ namespace CxxTest
         void failedAssertSameData( const char * /*file*/, unsigned /*line*/,
                                    const char * /*xStr*/, const char * /*yStr*/,
                                    const char * /*sizeStr*/, const void * /*x*/,
-                                   const void * /*y*/, unsigned /*size*/ )
+                                   const void * /*y*/, size_t /*size*/ )
         {
             redBarSafe();
         }
